# Specialize `TrustedLen` for `Iterator::unzip()`