@ChayimFriedman2 2024-07-07
@rustbot ready

Did the requested changes.