@the8472 2024-07-07
unzip is used in a few places in the compiler, so it might affect perf results

@bors r+ rollup=never